+++ BoardSquare.h
@@ -10,6 +10,11 @@ using std::to_string;
 #include <fstream>
 using std::ofstream;
 
+#define DEFAULT "\033[0m"
+#define REDTEXT "\033[31m"
+#define GREENTEXT "\033[32m"
+#define BLUEBOLDTEXT "\033[1;34m"
+
 class BoardSquare {
     private:
     enum class SquareStatus : uint8_t {preloaded, testing, empty};
@@ -38,6 +43,26 @@ class BoardSquare {
     bool occupied() const;
 
     friend inline ostream& operator<<(ostream &os, const BoardSquare &sq) {
-        return os << ((sq.value != 0) ? to_string(sq.value) : "_");
+        switch(sq.status) {
+            case SquareStatus::preloaded:
+                return os << BLUEBOLDTEXT << sq.value << DEFAULT;
+            break;
+
+            case SquareStatus::testing:
+                return os << GREENTEXT << sq.value << DEFAULT;
+            break;
+
+            case SquareStatus::empty:
+                return os << REDTEXT << "-" << DEFAULT;
+            break;
+        }
+    }
+
+    void inline print(ofstream &of) {
+        if(value == 0) {
+            of << "-";
+        } else {
+            of << value;
+        }
     }
 };