+++ main.cpp
@@ -9,6 +9,7 @@ using std::endl;
 using std::flush;
 #include <string>
 using std::string;
+#include <string.h>
 #include <fstream>
 using std::ifstream;
 #include <getopt.h>
@@ -18,9 +19,10 @@ using std::ifstream;
 ifstream inputFile;
 ofstream outputFile;
 
-bool solve = false;
+string solve;
+string generate;
 
-
+SudokuBoard::Difficulty difficulty = SudokuBoard::Difficulty::MEDIUM;
 
 void printBanner() {
     cout << "************************************************************\n*                          SUDOKU                          *\n*                Solving, Playing, and More                *\n*                                                          *\n*                A program by By Zalan Shah                *\n************************************************************\n"
@@ -36,14 +38,18 @@ void printHelp() {
     << "\t--help, -t\n"
     << "\t\t- This command will provide you with information on how to use this program, including all flags and options\n"
     << "\t--solve <filename>, -s <filename>\n"
-    << "\t\t- Takes a .txt file that contains an unsolved 9x9 board, and prints the solved board\n"
+    << "\t\t- Solves a given board in <filename>, and prints the solved board to the terminal\n"
     << "\t--play, -p\n"
     << "\t\t- This command is a work in progress :)\n"
     << "\t--generate [filename], -g [filename]\n"
-    << "\t\t- This command is a work in progreess :)\n"
+    << "\t\t- Generates a random sudoku board, and prints it to [filename] if given. If [filename] isn't given, the board is printed to the terminal\n"
+    << "\t\t- If the difficulty flag is being used, a board will be generated according to the difficulty specified.\n"
+    << "\t\t- If one isn't specified, the default difficulty is set to medium\n"
+    << "\t--difficulty <EASY|MEDIUM|HARD|IMPOSSIBLE>, -d <EASY|MEDIUM|HARD|IMPOSSIBLE>\n"
+    << "\t\t- Specifies the difficulty of a randomly generated board, and erases as many board squares as needed\n"
     << "\t--version, -v\n"
     << "\t\t- Prints the current version of this program\n\n"
-    << "Running without any flags will provide you with a menu to make decisions from" << endl;
+    << "Running without any flags will provide you with a menu to make decisions from." << endl;
 }
 
 
@@ -54,6 +60,7 @@ void clear() {
 }
 
 void printMenu() {
+    clear();
     printBanner();
 
     cout << "Welcome to Sudoku. Please select an option below to begin:\n"
@@ -80,14 +87,12 @@ void printMenu() {
             if(!inputFile.is_open()) {
                 cerr << "Error: File does not exist: " << input << endl;
                 sleep(2);
-                clear();
                 printMenu();
                 return;
             } else if(input.substr(input.find('.'), input.size()) != ".txt") {
                 cerr << "Error: Input must be a text file: " << input << endl;
                 inputFile.close();
                 sleep(2);
-                clear();
                 printMenu();
                 return;
             } else {
@@ -95,57 +100,52 @@ void printMenu() {
                 SudokuSolver solver(board);
 
                 if(solver.solveBoard()) {
-                    solver.printBoard();
+                    board.print();
                 } else {
                     cerr << "Error: The given board has no valid solution" << endl;
                 }
                 cout << "Press Enter to continue" << endl;
                 inputFile.close();
                 cin.get();
-                clear();
                 printMenu();
                 return;
             }
         break;
 
         case 'G':
-        cout << "This is currently a work in progress. Press Enter to continue." << endl;
-            // if(input.length() == 1) {
-            //     cout << "Enter output file name: " << flush;
-            //     getline(cin, input);
-            // } else {
-            //     input = input.substr(2, input.size());
-            // }
-            // outputFile.open(input);
+            if(input.length() == 1) {
+                cout << "Enter output file name: " << flush;
+                getline(cin, input);
+            } else {
+                input = input.substr(2, input.size());
+            }
+            outputFile.open(input);
 
-            // if(!outputFile.is_open()) {
-            //     cerr << "Error: File does not exist: " << input << endl;
-            //     sleep(2);
-            //     clear();
-            //     return printMenu();
-            // } else if(input.substr(input.find('.'), input.size()) != ".txt") {
-            //     cerr << "Error: Input must be a text file: " << input << endl;
-            //     outputFile.close();
-            //     sleep(2);
-            //     clear();
-            //     return printMenu();
-            // } else {
-            //     SudokuBoard board;
-            //     board.generate();
-            //     board.print(outputFile);
-
-            //     cout << "Generated output to " << input << ".\nPress Enter to continue" << endl;
-            //     outputFile.close();
-                cin.get();
+            if(!outputFile.is_open()) {
+                cerr << "Error: File does not exist: " << input << endl;
+                sleep(2);
                 clear();
                 return printMenu();
-            // }
+            } else if(input.substr(input.find('.'), input.size()) != ".txt") {
+                cerr << "Error: Input must be a text file: " << input << endl;
+                outputFile.close();
+                sleep(2);
+                clear();
+                return printMenu();
+            } else {
+                SudokuBoard board;
+                board.generate(difficulty);
+                board.print(outputFile);
+
+                cout << "Generated output to " << input << ".\nPress Enter to continue" << endl;
+                outputFile.close();
+                return printMenu();
+            }
 
         break;
         case 'P':
         cout << "This is currently a work in progress. Press Enter to continue." << endl;
         cin.get();
-        clear();
         return printMenu();
 
         break;
@@ -165,7 +165,6 @@ void printMenu() {
         cerr << "Error: Invalid option: " << input
         << "\nPress anywhere to continue" << flush;
         cin.get();
-        clear();
         printMenu();
         return;
         break;
@@ -176,16 +175,17 @@ void readArguments(int argc, char** argv) {
     int choice;
     int option_index = 0;
     option long_options[] = {
-        { "help",        no_argument,       nullptr, 'h'  },
-        { "solve",       required_argument, nullptr, 's'  },
-        { "play",        no_argument,       nullptr, 'p'  },
-        { "generate",    required_argument, nullptr, 'g'  },
-        { "version",     no_argument,       nullptr, 'v'  },
-        { nullptr,       no_argument,       nullptr, '\0' }
+        { "help",       no_argument,       nullptr, 'h'  },
+        { "solve",      required_argument, nullptr, 's'  },
+        { "play",       no_argument,       nullptr, 'p'  },
+        { "generate",   required_argument, nullptr, 'g'  },
+        { "difficulty", required_argument, nullptr, 'd'  },
+        { "version",    no_argument,       nullptr, 'v'  },
+        { nullptr,      no_argument,       nullptr, '\0' }
     };
 
     // while((choice = getopt_long(argc, argv, "hi:o:", long_options, &option_index)) != -1) {
-    while((choice = getopt_long(argc, argv, "hs:pg:v", long_options, &option_index)) != -1) {
+    while((choice = getopt_long(argc, argv, "hs:pg:d:v", long_options, &option_index)) != -1) {
         switch(choice) {
             case 'h': {
                 printHelp();
@@ -193,18 +193,17 @@ void readArguments(int argc, char** argv) {
             }
             break;
 
-            case 's': {
-                string fileName = optarg;
-                inputFile.open(fileName);
+            case 's':
+            solve = optarg;
+            inputFile.open(solve);
 
-                if(!inputFile.is_open()) {
-                    cerr << "Error: File does not exist: " << fileName << endl;
-                    exit(1);
-                } else if(fileName.substr(fileName.find('.'), fileName.size()) != ".txt") {
-                    cerr << "Error: Input must be a text file: " << fileName << endl;
-                    exit(1);
-                }
-                solve = true;
+            if(!inputFile.is_open()) {
+                cerr << "Error: File does not exist: " << solve << endl;
+                exit(1);
+            } else if(solve.substr(solve.find('.'), solve.size()) != ".txt") {
+                cerr << "Error: Input file must be a text file: " << solve << endl;
+                inputFile.close();
+                exit(1);
             }
             break;
 
@@ -214,14 +213,37 @@ void readArguments(int argc, char** argv) {
             }
             break;
 
-            case 'g': {
-                cout << "This is a work in progress :)" << endl;
-                exit(0);
+            case 'g':
+            generate = optarg;
+            outputFile.open(generate);
+
+            if(!outputFile.is_open()) {
+                cerr << "Error: File does not exist: " << generate << endl;
+                exit(1);
+            } else if(generate.substr(generate.find('.'), generate.size()) != ".txt") {
+                cerr << "Error: Output file must be a text file: " << generate << endl;
+                outputFile.close();
+                exit(1);
+            }
+            break;
+
+            case 'd':
+            if(strcmp(optarg, "EASY") == 0 || strcmp(optarg, "easy") == 0) {
+                difficulty = SudokuBoard::Difficulty::EASY;
+            } else if(strcmp(optarg, "MEDIUM") == 0 || strcmp(optarg, "medium") == 0) {
+                difficulty = SudokuBoard::Difficulty::MEDIUM;
+            } else if(strcmp(optarg, "HARD") == 0 || strcmp(optarg, "hard") == 0) {
+                difficulty = SudokuBoard::Difficulty::HARD;
+            } else if(strcmp(optarg, "IMPOSSIBLE") == 0 || strcmp(optarg, "impossible") == 0) {
+                difficulty = SudokuBoard::Difficulty::IMPOSSIBLE;
+            } else {
+                cerr << "Error: Unknown difficulty: " << optarg << ". Difficulty may only be EASY, MEDIUM, HARD, or IMPOSSIBLE" << endl;
+                exit(1);
             }
             break;
 
             case 'v': {
-                cout << "v1.0.0" << endl;
+                cout << "v1.1.0" << endl;
                 exit(0);
             }
         }
@@ -231,29 +253,25 @@ void readArguments(int argc, char** argv) {
 int main(int argc, char **argv) {
     readArguments(argc, argv);
 
-    if(system("clear") != 0) {
-        exit(1);
-    }
-
-
-
-    if(solve) {
+    if(solve != "") {
         SudokuBoard board(inputFile);
+        inputFile.close();
         SudokuSolver solver(board);
 
         if(solver.solveBoard()) {
-            solver.printBoard();
+            board.print();
             exit(0);
         } else {
             cerr << "Error: The given board has no valid solution" << endl;
             exit(1);
         }
+    } else if(generate != "") {
+        SudokuBoard board;
+        board.generate(difficulty);
+
+        board.print(outputFile);
+        exit(0);
     }
-    clear();
+
     printMenu();
-
-
-
-
-    // SudokuSolver sudoku(argc, argv);
 }