+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v1.1.0: Generate, Difficulties, Splash of Color
+
+### Added or Changed
+- Added Generate feature
+    - Added options for difficulty when generating board
+- Added some colors to distinguish between "given" and "solved" squares when solving a board
+- Updated --help flag
+- Updated Readme
+- Fixed bug where ./sudoku -s \<`filename`\> clears terminal
+
 ## v1.0.0: First implementation
 
 ### Added or Changed